separates signinbuttons into 2 components
