+++ src/applications/login/components/FedWarning.jsx
@@ -1,15 +1,15 @@
 import React from 'react';
 
-export default function FedWarning({ newDesignEnabled }) {
+export default function FedWarning({ oldDesignEnabled }) {
   return (
     <div
       className={`${
-        newDesignEnabled
+        !oldDesignEnabled
           ? 'vads-u-padding-bottom--2p5 fed-warning--v2'
           : 'fed-warning'
       }`}
     >
-      {newDesignEnabled && (
+      {!oldDesignEnabled && (
         <h2 className="vads-u-margin-top--0">Terms of use</h2>
       )}
       <p>
