+++ src/platform/user/authentication/components/SignInModal.jsx
@@ -5,7 +5,8 @@ import { connect } from 'react-redux';
 import Modal from '@department-of-veterans-affairs/component-library/Modal';
 
 import FedWarning from 'applications/login/components/FedWarning';
-import SignInButtons from 'applications/login/components/SignInButtons';
+import NewDesignButtons from 'applications/login/components/NewDesignButtons';
+import OriginalDesignButtons from 'applications/login/components/OriginalDesignButtons';
 import SubmitSignInForm from 'platform/static-data/SubmitSignInForm';
 import SignInDescription from 'applications/login/components/SignInDescription';
 
@@ -17,7 +18,7 @@ import {
   ssoe,
   loginGov,
   loginGovCreateAccount,
-  loginNewDesign,
+  loginOldDesign,
 } from 'platform/user/authentication/selectors';
 import { login, signup } from 'platform/user/authentication/utilities';
 import { formatDowntime } from 'platform/utilities/date';
@@ -81,65 +82,7 @@ export class SignInModal extends React.Component {
                 Sign in with an existing account
               </h2>
               <div>
-                <button
-                  disabled={globalDowntime}
-                  className="dslogon"
-                  onClick={this.loginHandler('dslogon')}
-                >
-                  <img
-                    aria-hidden="true"
-                    role="presentation"
-                    alt="DS Logon"
-                    src={`${vaGovFullDomain}/img/signin/dslogon-icon.svg`}
-                  />
-                  <strong> Sign in with DS Logon</strong>
-                </button>
-                <button
-                  disabled={globalDowntime}
-                  className="mhv"
-                  onClick={this.loginHandler('mhv')}
-                >
-                  <img
-                    aria-hidden="true"
-                    role="presentation"
-                    alt="My HealtheVet"
-                    src={`${vaGovFullDomain}/img/signin/mhv-icon.svg`}
-                  />
-                  <strong> Sign in with My HealtheVet</strong>
-                </button>
-                <button
-                  disabled={globalDowntime}
-                  className="usa-button-primary va-button-primary"
-                  onClick={this.loginHandler('idme')}
-                >
-                  <img
-                    aria-hidden="true"
-                    role="presentation"
-                    alt="ID.me"
-                    src={`${vaGovFullDomain}/img/signin/idme-icon-white.svg`}
-                  />
-                  <strong> Sign in with ID.me</strong>
-                </button>
-                <span className="sidelines">OR</span>
-                <div className="alternate-signin">
-                  <h2 className="vads-u-font-size--sm vads-u-margin-top--0">
-                    Don't have those accounts?
-                  </h2>
-                  <button
-                    disabled={globalDowntime}
-                    className="idme-create usa-button usa-button-secondary"
-                    onClick={this.signupHandler}
-                  >
-                    <img
-                      aria-hidden="true"
-                      role="presentation"
-                      alt="ID.me"
-                      src={`${vaGovFullDomain}/img/signin/idme-icon-dark.svg`}
-                    />
-                    <strong> Create an ID.me account</strong>
-                  </button>
-                  <p>Use your email, Google, or Facebook</p>
-                </div>
+                <OriginalDesignButtons isDisabled={globalDowntime} />
               </div>
             </div>
           </div>
@@ -247,14 +190,14 @@ export class SignInModal extends React.Component {
         <div className="row">
           <div className="columns small-12">
             <h1
-              className={`${this.props.newDesignEnabled &&
+              className={`${!this.props.oldDesignEnabled &&
                 'vads-u-margin-top--2 medium-screen:vads-u-margin-top--1 medium-screen:vads-u-margin-bottom--2'}`}
             >
               Sign in
             </h1>
           </div>
         </div>
-        {!this.props.newDesignEnabled ? (
+        {this.props.oldDesignEnabled ? (
           <div className="row medium-screen:vads-u-display--none mobile-explanation">
             <div className="columns small-12">
               <h2>
@@ -265,18 +208,17 @@ export class SignInModal extends React.Component {
           </div>
         ) : null}
         {this.renderDowntimeBanners()}
-        {!this.props.newDesignEnabled ? (
+        {this.props.oldDesignEnabled ? (
           this.renderOriginalModal({
             globalDowntime,
           })
         ) : (
           <div className="row">
-            <SignInButtons
+            <NewDesignButtons
               loginGovEnabled={this.props.loginGovEnabled}
               loginGovCreateAccountEnabled={
                 this.props.loginGovCreateAccountEnabled
               }
-              newDesignEnabled={this.props.newDesignEnabled}
               isDisabled={globalDowntime}
             />
           </div>
@@ -286,7 +228,7 @@ export class SignInModal extends React.Component {
             <div className="help-info">
               <h2
                 className={`${
-                  this.props.newDesignEnabled
+                  !this.props.oldDesignEnabled
                     ? 'vads-u-margin-top--0'
                     : 'vads-u-font-size--md'
                 }`}
@@ -314,10 +256,10 @@ export class SignInModal extends React.Component {
               </p>
               <p>
                 <SubmitSignInForm startSentence />{' '}
-                {this.props.newDesignEnabled && ` We're here 24/7.`}
+                {!this.props.oldDesignEnabled && ` We're here 24/7.`}
               </p>
             </div>
-            <FedWarning newDesignEnabled={this.props.newDesignEnabled} />
+            <FedWarning oldDesignEnabled={this.props.oldDesignEnabled} />
           </div>
         </div>
       </div>
@@ -328,7 +270,7 @@ export class SignInModal extends React.Component {
     return (
       <Modal
         cssClass={`va-modal-large ${
-          this.props.newDesignEnabled ? 'new-modal-design' : ''
+          !this.props.oldDesignEnabled ? 'new-modal-design' : ''
         }`}
         visible={this.props.visible}
         focusSelector="button"
@@ -351,7 +293,7 @@ function mapStateToProps(state) {
     useSSOe: ssoe(state),
     loginGovEnabled: loginGov(state),
     loginGovCreateAccountEnabled: loginGovCreateAccount(state),
-    newDesignEnabled: loginNewDesign(state),
+    oldDesignEnabled: loginOldDesign(state),
   };
 }
